+++ pubspec.yaml
@@ -20,7 +20,7 @@ dev_dependencies:
 
 flutter:
   assets:
-    - assets/
+    - assets/app-logo.png
   uses-material-design: true
